+++ git_issue_reporter/core.py
